# src/main/resources/com/sonicle/webtop/drm/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop DRM",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is DRM service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    },
    {
      "auth-apikey-bearer": [],
      "auth-apikey-username": []
    }
  ],
  "paths": {
    "/timetable": {
      "get": {
        "summary": "List Timetable Entries",
        "operationId": "listTimetableEntries",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableEntriesResult"
                }
              }
            }
          }
        },
        "description": "List timetable entries for a specific month / year",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "operatorId",
            "description": "Effective user"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "year"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "month"
          }
        ],
        "tags": [
          "me"
        ]
      },
      "post": {
        "summary": "Punch",
        "operationId": "punch",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "type",
            "description": "M = Manual , S = Special"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "location",
            "description": "O = Office, S = Smart, A = App"
          }
        ],
        "description": "Register a Punch (Clock-in / Clock-out automatic)"
      }
    },
    "/leaveRequests": {
      "get": {
        "summary": "List Leave Requests",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestsResult"
                }
              }
            }
          }
        },
        "operationId": "listLeaveRequests",
        "description": "List Leave Requests",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "userId"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "startDate"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "leaveRequestId"
          }
        ]
      },
      "parameters": [],
      "post": {
        "summary": "Add new Leave Requesst",
        "operationId": "addLeaveRequest",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Add a new Leave Request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequest"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Leave Request",
        "operationId": "updateLeaveRequests",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Update Leave Request",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequest"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Leave Request",
        "operationId": "deleteLeaveRequest",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Delete a Leave Request",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "leaveRequestId"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "delete",
                "approve",
                "refuse",
                "cancel",
                "requestCancellation"
              ]
            },
            "in": "query",
            "name": "action"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "text"
          }
        ]
      }
    },
    "/managers": {
      "get": {
        "summary": "List Managers",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagersResult"
                }
              }
            }
          }
        },
        "operationId": "listManagers",
        "description": "List Managers of a specific user",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "userId"
          }
        ]
      }
    },
    "/companies": {
      "get": {
        "summary": "List Companies",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompaniesResult"
                }
              }
            }
          }
        },
        "operationId": "listCompanies",
        "description": "List Companies of a specific user",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "userId"
          }
        ]
      },
      "parameters": []
    },
    "/employee": {
      "get": {
        "summary": "Get Employee Profile",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          }
        },
        "operationId": "getEmployeeProfile",
        "description": "Get Employee Profile data"
      }
    },
    "/managers/users": {
      "get": {
        "summary": "List managed Users by a Manager",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResult"
                }
              }
            }
          }
        },
        "operationId": "listManagedUsers",
        "description": "List managed Users by a Manager",
        "parameters": []
      }
    }
  },
  "components": {
    "schemas": {
      "TimetableEntry": {
        "title": "TimetableEntry",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "x-stoplight": {
              "id": "tjgxq0x1jwjga"
            }
          },
          "date": {
            "type": "string",
            "x-stoplight": {
              "id": "gxcdxiy42683e"
            }
          },
          "entrance": {
            "type": "string",
            "description": "Entrance date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "exit": {
            "type": "string",
            "description": "Exit date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "location": {
            "type": "string",
            "description": "Exit date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "minutes": {
            "type": "integer",
            "x-stoplight": {
              "id": "p6cw9o2xt6v1z"
            }
          }
        },
        "required": [
          "entrance",
          "exit",
          "location"
        ]
      },
      "ApiError": {
        "title": "ApiError",
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          }
        },
        "description": "This general detailed error structure is used throughout this API."
      },
      "TimetableEntriesResult": {
        "title": "TimetableEntriesResult",
        "x-stoplight": {
          "id": "mss129ka16usn"
        },
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableEntry"
            }
          }
        }
      },
      "LeaveRequest": {
        "title": "LeaveRequest",
        "x-stoplight": {
          "id": "57xpe5uaa4w7h"
        },
        "type": "object",
        "properties": {
          "leaveRequestId": {
            "type": "integer",
            "x-stoplight": {
              "id": "015hosf3elqtr"
            }
          },
          "domainId": {
            "type": "string",
            "x-stoplight": {
              "id": "9a931qf8bthip"
            }
          },
          "companyId": {
            "type": "integer",
            "x-stoplight": {
              "id": "mke22xscg7ejj"
            }
          },
          "userId": {
            "type": "string",
            "x-stoplight": {
              "id": "ts2f092apd7ak"
            }
          },
          "user": {
            "type": "string",
            "x-stoplight": {
              "id": "22s9rmjbtf2cu"
            }
          },
          "managerId": {
            "type": "string",
            "x-stoplight": {
              "id": "9xmygt62rwreo"
            }
          },
          "manager": {
            "type": "string",
            "x-stoplight": {
              "id": "f1z4nt60vmqtq"
            }
          },
          "type": {
            "type": "string",
            "x-stoplight": {
              "id": "acz7vdqozy2uo"
            }
          },
          "fromDate": {
            "type": "string",
            "x-stoplight": {
              "id": "w5vhzwg2goi8q"
            }
          },
          "toDate": {
            "type": "string",
            "x-stoplight": {
              "id": "lcz2uslkisazg"
            }
          },
          "fromHour": {
            "type": "string",
            "x-stoplight": {
              "id": "xnuqbd5sqccv4"
            }
          },
          "toHour": {
            "type": "string",
            "x-stoplight": {
              "id": "zasae7pl9redn"
            }
          },
          "status": {
            "type": "string",
            "x-stoplight": {
              "id": "62kyq18aqt8sm"
            }
          },
          "result": {
            "type": "boolean",
            "x-stoplight": {
              "id": "wdpe3cyz1rc7c"
            }
          },
          "employeeCancReq": {
            "type": "boolean",
            "x-stoplight": {
              "id": "o0if6mtb86tow"
            }
          },
          "notes": {
            "type": "string",
            "x-stoplight": {
              "id": "dextkc0tou574"
            }
          },
          "cancReason": {
            "type": "string",
            "x-stoplight": {
              "id": "hp5sfutpoa75x"
            }
          }
        }
      },
      "LeaveRequestsResult": {
        "title": "LeaveRequestsResult",
        "x-stoplight": {
          "id": "by5oddjk4o9ip"
        },
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "x-stoplight": {
              "id": "h104bx2n7l37q"
            },
            "items": {
              "$ref": "#/components/schemas/LeaveRequest"
            }
          }
        }
      },
      "Manager": {
        "title": "Manager",
        "x-stoplight": {
          "id": "yh9bpqqt9b35j"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayname": {
            "type": "string",
            "x-stoplight": {
              "id": "s4rewoog9tnfq"
            }
          }
        }
      },
      "ManagersResult": {
        "title": "ManagersResult",
        "x-stoplight": {
          "id": "bqxurdff8em6c"
        },
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "x-stoplight": {
              "id": "ee7uunyhguq24"
            },
            "items": {
              "$ref": "#/components/schemas/Manager"
            }
          }
        }
      },
      "Company": {
        "title": "Company",
        "x-stoplight": {
          "id": "uwm6t998n9w6q"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayname": {
            "type": "string",
            "x-stoplight": {
              "id": "c02dzsrtgshty"
            }
          }
        }
      },
      "CompaniesResult": {
        "title": "CompaniesResult",
        "x-stoplight": {
          "id": "pn3k81o0yte29"
        },
        "type": "object",
        "properties": {
          "items": {
            "x-stoplight": {
              "id": "sd6bvesvhxn48"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Company"
            }
          }
        }
      },
      "Employee": {
        "title": "Employee",
        "x-stoplight": {
          "id": "xkd2308nud21h"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "domainId": {
            "type": "string",
            "x-stoplight": {
              "id": "tluvpfwfluaqw"
            }
          },
          "userId": {
            "type": "string",
            "x-stoplight": {
              "id": "h3cct9i63o82k"
            }
          },
          "number": {
            "type": "string",
            "x-stoplight": {
              "id": "ede4eemun6ehm"
            }
          },
          "tolerance": {
            "type": "string",
            "x-stoplight": {
              "id": "a1r51fg8kn09o"
            }
          },
          "extraordinary": {
            "type": "boolean",
            "x-stoplight": {
              "id": "xtp1cfi329y3l"
            }
          },
          "onlyPresence": {
            "type": "boolean",
            "x-stoplight": {
              "id": "qlgacanijj7py"
            }
          },
          "hourProfileId": {
            "type": "integer",
            "x-stoplight": {
              "id": "wvvlfp6v4ulbw"
            }
          },
          "headquartersCode": {
            "type": "string",
            "x-stoplight": {
              "id": "r4c6rryatkonm"
            }
          },
          "noStamping": {
            "type": "boolean",
            "x-stoplight": {
              "id": "0kos2srgduhc9"
            }
          },
          "minimumNumberOfHoursPerTicket": {
            "type": "integer",
            "x-stoplight": {
              "id": "1tbpve3nwm7hd"
            }
          },
          "stampingMode": {
            "type": "string",
            "x-stoplight": {
              "id": "9gzegu8639ei0"
            }
          },
          "isManager": {
            "type": "boolean",
            "x-stoplight": {
              "id": "8vc4d2grc8bkg"
            }
          },
          "leaveRequestTypes": {
            "type": "array",
            "x-stoplight": {
              "id": "bpm7lzrukfxl5"
            },
            "items": {
              "x-stoplight": {
                "id": "zraxp0d5vyhh8"
              },
              "type": "string"
            }
          }
        }
      },
      "User": {
        "title": "User",
        "x-stoplight": {
          "id": "bbjj6mzldor9r"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayname": {
            "type": "string",
            "x-stoplight": {
              "id": "zkreiwi1heqa9"
            }
          }
        }
      },
      "UsersResult": {
        "title": "UsersResult",
        "x-stoplight": {
          "id": "1g10mhxoix3t1"
        },
        "type": "object",
        "properties": {
          "items": {
            "x-stoplight": {
              "id": "p53volqj3tip5"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "auth-apikey-bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "APIKey",
        "description": "API-Key Authentication: specifies the secret token (bearer and username are both required)"
      },
      "auth-apikey-username": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Username",
        "description": "API-Key Authentication: specifies target username (bearer and username are both required)"
      }
    }
  },
  "x-original-swagger-version": "2.0"
}